Fix ramp exit
When going down a ramp there would be a sudden small jump. This was
due to jump building up going down the ramp and the player starting to "jump"
just before leaving the ramp as the player can be consider to be "off"
the ramp on the last frame before exiting. The later still happens but
the time in the air is much smaller at lower speeds and still a thing
at higher speeds but that doesn't feel problematic when playing (at
this point at least).