Tweak player movement
Communicate the intertia of moving up and down a bit better and have a
lower limit on the amount of thrust so small taps generally have the
desired effect. Also moved the RHS limit to the centre of the screen
as there's no benefit to being further right and a fun challenge
trying not to crash at high speed.