Garbage collect snow
A few optimizations to make it run better for the number of snowflakes
that have landed both in terms of speed (calling spr() * 1000 reduces
the frame rate to 30FPS!) and memory (around 5000 snow flakes with
coroutines attached would make the cart run out of memory). Still not
perfect but it's slightly less fragile now.